requirements.txt
pandas==2.1.1
requests==2.31.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
openpyxl==3.1.2
minio==7.1.3
python-dotenv
psycopg2-binary
beautifulsoup4==4.12.2
